# wc: Speed optimization