@jnsgruk 2025-05-15
Nice, same result on my machine.

```
❯ hyperfine --warmup 3 "wc -l ~/temp/measurements.txt" "target/release/wc -l ~/temp/measurements.txt"
Benchmark 1: wc -l ~/omgrust/measurements.txt
  Time (mean ± σ):      1.528 s ±  0.068 s    [User: 0.166 s, System: 1.362 s]
  Range (min … max):    1.435 s …  1.626 s    10 runs

Benchmark 2: target/release/wc -l ~/omgrust/measurements.txt
  Time (mean ± σ):      1.378 s ±  0.074 s    [User: 0.182 s, System: 1.195 s]
  Range (min … max):    1.270 s …  1.525 s    10 runs

Summary
  target/release/wc -l ~/temp/measurements.txt ran
    1.11 ± 0.08 times faster than wc -l ~/temp/measurements.txt
```