@drinkcat 2025-05-15
> Nice, same result on my machine, which is a 7040 series Ryzen laptop chip.

It's actually a lot better than what I see. Interesting! Thanks for testing.