@willshuttleworth 2025-05-15
previously the coreutils implementation was about 10x faster on my m1 mac, now its about 13x faster!
```
hyperfine -w 3 'wc -l /tmp/seq100M' './target/release/wc -l /tmp/seq100M'
Benchmark 1: wc -l /tmp/seq100M
  Time (mean ± σ):     900.0 ms ±   0.3 ms    [User: 844.2 ms, System: 55.4 ms]
  Range (min … max):   899.6 ms … 900.4 ms    10 runs

Benchmark 2: ./target/release/wc -l /tmp/seq100M
  Time (mean ± σ):      66.8 ms ±   0.8 ms    [User: 12.6 ms, System: 53.8 ms]
  Range (min … max):    65.8 ms …  70.6 ms    43 runs

Summary
  ./target/release/wc -l /tmp/seq100M ran
   13.47 ± 0.17 times faster than wc -l /tmp/seq100M
  ```