@sylvestre 2025-05-15
> Summary
>   ./target/release/wc -l /tmp/seq100M ran
>    13.47 ± 0.17 times faster than wc -l /tmp/seq100M
> ```

I wonder how it is possible :)

wc is the BSD apple implementation ?