@drinkcat 2025-05-15
@willshuttleworth oh nice, thanks!

I'm not completely sure about the aarch64 code, but I see some loading operations on 8x16x4, which sound like 512-bit/64-bytes (the underlying registers are 128-bit wide though).

Mind trying to see if 64-byte alignment improves performance?
```
diff --git a/src/uu/wc/src/count_fast.rs b/src/uu/wc/src/count_fast.rs
index 479183263b21..0dd7c3882a31 100644
--- a/src/uu/wc/src/count_fast.rs
+++ b/src/uu/wc/src/count_fast.rs
@@ -201,7 +201,7 @@ pub(crate) fn count_bytes_fast<T: WordCountable>(handle: &mut T) -> (usize, Opti
 ///
 /// This is useful as bytecount uses 256-bit wide vector operations that run much
 /// faster on aligned data (at least on x86 with AVX2 support).
-#[repr(align(32))]
+#[repr(align(64))]
 struct AlignedBuffer {
     data: [u8; BUF_SIZE],
 }
```
(you can `mv target/release/wc wc.main` before the change to easily compare)

Thanks!